backend/requirements.txt
transformers
diffusers
groq
requests
bs4
python-dotenv
gradio
pymongo
schedule
Pillow